+++ tests/unit/fem/test_face_permutation.cpp
@@ -167,7 +167,7 @@ double TestFaceRestriction(Mesh &mesh, int order)
    L2FaceRestriction restr(fes, ElementDofOrdering::LEXICOGRAPHIC,
                            FaceType::Interior, L2FaceValues::DoubleValued);
 
-   const int ndof_face = pow(order+1, dim-1);
+   const int ndof_face = static_cast<int>(pow(order+1, dim-1));
 
    Vector face_values(ndof_face*2);
